requirements.txt
python-dotenv
fastapi 
requests
PyPDF2
azure-storage-blob
azure-identity
azure-search-documents
azure-cognitiveservices-speech
openai
google-generativeai
boto3
pandas
#azure-ai-vision==0.15.1b1
#azure-ai-vision-imageanalysis
uvicorn
pytest 
pytest-asyncio 
httpx
azure-identity 
azure-keyvault-secrets
pydantic_settings
pillow
python-multipart
pyautogen
markdownify
scikit-learn
xgboost
autogen-agentchat
mammoth
pdfminer
pdfminer-six
python-pptx
puremagic
pathvalidate
pyodbc
agentops
autogen-ext==0.4.0.dev9
autogen-agentchat==0.4.0.dev9
autogen-core==0.4.0.dev9
markitdown
matplotlib
yfinance
seaborn
tabulate